+++ crates/pallets/starknet/src/lib.rs
@@ -33,7 +33,6 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 #![allow(clippy::large_enum_variant)]
 
-use blockifier::transaction::objects::TransactionExecutionResult;
 /// Starknet pallet.
 /// Definition of the pallet's runtime storage items, events, errors, and dispatchable
 /// functions.
@@ -46,19 +45,15 @@ pub mod blockifier_state_adapter;
 pub mod execution_config;
 #[cfg(feature = "std")]
 pub mod genesis_loader;
+/// Simulation, estimations and execution trace logic.
+pub mod simulations;
 /// Transaction validation logic.
 pub mod transaction_validation;
 /// The Starknet pallet's runtime custom types.
 pub mod types;
 
-use blockifier::execution::entry_point::{CallEntryPoint, CallType, EntryPointExecutionContext};
-use blockifier::state::cached_state::ContractStorageKey;
-use starknet_api::state::StorageKey;
-use starknet_api::transaction::{Calldata, Event as StarknetEvent, Fee};
-
 #[cfg(test)]
 mod tests;
-mod utils;
 
 #[macro_use]
 pub extern crate alloc;
@@ -71,8 +66,11 @@ use alloc::vec::Vec;
 
 use blockifier::block_context::BlockContext;
 use blockifier::execution::contract_class::ContractClass;
-use blockifier::execution::entry_point::{CallInfo, ExecutionResources};
+use blockifier::execution::entry_point::{
+    CallEntryPoint, CallInfo, CallType, EntryPointExecutionContext, ExecutionResources,
+};
 use blockifier::execution::errors::{EntryPointExecutionError, PreExecutionError};
+use blockifier::state::cached_state::ContractStorageKey;
 use blockifier_state_adapter::BlockifierStateAdapter;
 use frame_support::pallet_prelude::*;
 use frame_support::traits::Time;
@@ -85,11 +83,6 @@ use mp_fee::{ResourcePrice, INITIAL_GAS};
 use mp_felt::Felt252Wrapper;
 use mp_hashers::HasherT;
 use mp_sequencer_address::{InherentError, InherentType, DEFAULT_SEQUENCER_ADDRESS, INHERENT_IDENTIFIER};
-use mp_simulations::{
-    DeclareTransactionTrace, DeployAccountTransactionTrace, FeeEstimate, FunctionInvocation, InvokeTransactionTrace,
-    SimulatedTransaction, SimulationFlags, TransactionTrace,
-};
-use mp_state::rpc::StateDiff;
 use mp_storage::{StarknetStorageSchemaVersion, PALLET_STARKNET_SCHEMA};
 use mp_transactions::execution::Execute;
 use mp_transactions::{
@@ -101,20 +94,22 @@ use sp_runtime::DigestItem;
 use starknet_api::api_core::{ChainId, ClassHash, CompiledClassHash, ContractAddress, EntryPointSelector, Nonce};
 use starknet_api::block::{BlockNumber, BlockTimestamp};
 use starknet_api::deprecated_contract_class::EntryPointType;
-use starknet_api::hash::StarkFelt;
-use starknet_api::transaction::{MessageToL1, TransactionHash};
+use starknet_api::hash::{StarkFelt, StarkHash};
+use starknet_api::state::StorageKey;
+use starknet_api::transaction::{Calldata, Event as StarknetEvent, Fee, MessageToL1, TransactionHash};
 use starknet_crypto::FieldElement;
 use transaction_validation::TxPriorityInfo;
 
 use crate::alloc::string::ToString;
 use crate::execution_config::RuntimeExecutionConfigBuilder;
 use crate::types::{CasmClassHash, SierraClassHash, StorageSlot};
-use crate::utils::{convert_call_info_to_execute_invocation, execute_txs_and_rollback};
 
 pub(crate) const LOG_TARGET: &str = "runtime::starknet";
 
 pub const ETHEREUM_EXECUTION_RPC: &[u8] = b"starknet::ETHEREUM_EXECUTION_RPC";
 pub const ETHEREUM_CONSENSUS_RPC: &[u8] = b"starknet::ETHEREUM_CONSENSUS_RPC";
+
+pub const SN_OS_CONFIG_HASH_VERSION: &str = "StarknetOsConfig1";
 
 // syntactic sugar for logging.
 #[macro_export]
@@ -145,8 +140,11 @@ pub mod pallet {
         type RuntimeEvent: From<Event<Self>> + IsType<<Self as frame_system::Config>::RuntimeEvent>;
         /// The hashing function to use.
         type SystemHash: HasherT;
-        /// The time idk what.
+        /// The block time
         type TimestampProvider: Time;
+        /// The gas price
+        #[pallet::constant]
+        type L1GasPrice: Get<ResourcePrice>;
         /// A configuration for base priority of unsigned transactions.
         ///
         /// This is exposed so that it can be tuned for particular runtime, when
@@ -169,7 +167,7 @@ pub mod pallet {
         #[pallet::constant]
         type ValidateMaxNSteps: Get<u32>;
         #[pallet::constant]
-        type ProtocolVersion: Get<u8>;
+        type ProtocolVersion: Get<Felt252Wrapper>;
         #[pallet::constant]
         type ChainId: Get<Felt252Wrapper>;
         #[pallet::constant]
@@ -437,14 +435,14 @@ pub mod pallet {
         /// second element is the contract class hash.
         /// This can be used to start the chain with a set of pre-deployed contracts, for example in
         /// a test environment or in the case of a migration of an existing chain state.
-        pub contracts: Vec<(ContractAddress, CasmClassHash)>,
+        pub contracts: Vec<(ContractAddress, SierraClassHash)>,
         pub sierra_to_casm_class_hash: Vec<(SierraClassHash, CasmClassHash)>,
         /// The contract classes to be deployed at genesis.
         /// This is a vector of tuples, where the first element is the contract class hash and the
         /// second element is the contract class definition.
         /// Same as `contracts`, this can be used to start the chain with a set of pre-deployed
         /// contracts classes.
-        pub contract_classes: Vec<(CasmClassHash, ContractClass)>,
+        pub contract_classes: Vec<(SierraClassHash, ContractClass)>,
         pub storage: Vec<(ContractStorageKey, StarkFelt)>,
         /// The address of the fee token.
         /// Must be set to the address of the fee token ERC20 contract.
@@ -539,8 +537,9 @@ pub mod pallet {
         Unimplemented,
         MissingRevertReason,
         MissingCallInfo,
-        TransactionalExecutionFailed,
+        FailedToCreateATransactionalStorageExecution,
         L1MessageAlreadyExecuted,
+        MissingL1GasUsage,
     }
 
     /// The Starknet pallet external functions.
@@ -918,8 +917,6 @@ impl<T: Config> Pallet<T> {
         let chain_id = Self::chain_id_str();
 
         let vm_resource_fee_cost = Default::default();
-        // FIXME: https://github.com/keep-starknet-strange/madara/issues/329
-        let gas_price = 10;
         BlockContext {
             block_number: BlockNumber(block_number),
             block_timestamp: BlockTimestamp(block_timestamp),
@@ -929,7 +926,7 @@ impl<T: Config> Pallet<T> {
             vm_resource_fee_cost,
             invoke_tx_max_n_steps: T::InvokeTxMaxNSteps::get(),
             validate_max_n_steps: T::ValidateMaxNSteps::get(),
-            gas_price,
+            gas_price: T::L1GasPrice::get().price_in_wei,
             max_recursion_depth: T::MaxRecursionDepth::get(),
         }
     }
@@ -1065,24 +1062,20 @@ impl<T: Config> Pallet<T> {
                 transaction_hashes.len(),
                 "transactions and transaction hashes should be the same length"
             );
+            let global_state_root = Felt252Wrapper::default();
             let transaction_count = transactions.len();
-
             let parent_block_hash = Self::parent_block_hash(&block_number);
             let events: Vec<StarknetEvent> = transaction_hashes.iter().flat_map(TxEvents::<T>::take).collect();
-
-            let global_state_root = Felt252Wrapper::default();
-
             let sequencer_address = Self::sequencer_address();
             let block_timestamp = Self::block_timestamp();
-
             let chain_id = Self::chain_id();
             let (transaction_commitment, event_commitment) =
                 mp_commitments::calculate_commitments::<T::SystemHash>(&transactions, &events, chain_id, block_number);
             let protocol_version = T::ProtocolVersion::get();
             let extra_data = None;
+            let l1_gas_price = T::L1GasPrice::get();
 
-            // TODO: Compute l1_gas_price correctly
-            let l1_gas_price = ResourcePrice::default();
+            let ordered_events = vec![];
 
             let block = StarknetBlock::new(
                 StarknetHeader::new(
@@ -1095,11 +1088,12 @@ impl<T: Config> Pallet<T> {
                     transaction_commitment.into(),
                     events.len() as u128,
                     event_commitment.into(),
-                    protocol_version,
+                    protocol_version.into(),
                     l1_gas_price,
                     extra_data,
                 ),
                 transactions,
+                ordered_events,
             );
             // Save the block number <> hash mapping.
             let blockhash = block.header().hash::<T::SystemHash>();
@@ -1211,132 +1205,6 @@ impl<T: Config> Pallet<T> {
         next_order
     }
 
-    /// Estimate the fee associated with transaction
-    pub fn estimate_fee(transactions: Vec<UserTransaction>) -> Result<Vec<(u64, u64)>, DispatchError> {
-        let chain_id = Self::chain_id();
-
-        let execution_results = execute_txs_and_rollback::<T>(
-            &transactions,
-            &Self::get_block_context(),
-            chain_id,
-            &mut RuntimeExecutionConfigBuilder::new::<T>().with_query_mode().build(),
-        )?;
-
-        let mut results = vec![];
-        for res in execution_results {
-            match res {
-                Ok(tx_exec_info) => {
-                    log!(info, "Successfully estimated fee: {:?}", tx_exec_info);
-                    if let Some(l1_gas_usage) = tx_exec_info.actual_resources.0.get("l1_gas_usage") {
-                        results.push((tx_exec_info.actual_fee.0 as u64, *l1_gas_usage as u64));
-                    } else {
-                        return Err(Error::<T>::TransactionExecutionFailed.into());
-                    }
-                }
-                Err(e) => {
-                    log!(info, "Failed to estimate fee: {:?}", e);
-                    return Err(Error::<T>::TransactionExecutionFailed.into());
-                }
-            }
-        }
-        Ok(results)
-    }
-
-    pub fn simulate_transactions(
-        transactions: Vec<UserTransaction>,
-        simulation_flags: SimulationFlags,
-    ) -> Result<Vec<SimulatedTransaction>, DispatchError> {
-        let chain_id = Self::chain_id();
-
-        let execution_results = execute_txs_and_rollback::<T>(
-            &transactions,
-            &Self::get_block_context(),
-            chain_id,
-            &mut RuntimeExecutionConfigBuilder::new::<T>().with_simulation_mode(&simulation_flags).build(),
-        )?;
-
-        fn get_function_invocation(
-            call_info: Option<&CallInfo>,
-        ) -> TransactionExecutionResult<Option<FunctionInvocation>> {
-            call_info.map(FunctionInvocation::try_from).transpose()
-        }
-
-        let mut results = vec![];
-        for (tx, res) in transactions.iter().zip(execution_results.iter()) {
-            match res {
-                Ok(tx_exec_info) => {
-                    let validate_invocation = get_function_invocation(tx_exec_info.validate_call_info.as_ref())
-                        .map_err(|err| {
-                            log::error!("Failed to convert validate call info to function invocation: {}", err);
-                            Error::<T>::TransactionExecutionFailed
-                        })?;
-                    let fee_transfer_invocation = get_function_invocation(tx_exec_info.fee_transfer_call_info.as_ref())
-                        .map_err(|err| {
-                            log::error!("Failed to convert fee transfer call info to function invocation: {}", err);
-                            Error::<T>::TransactionExecutionFailed
-                        })?;
-
-                    let transaction_trace = match tx {
-                        UserTransaction::Invoke(_) => {
-                            TransactionTrace::Invoke(InvokeTransactionTrace {
-                                validate_invocation,
-                                execute_invocation: convert_call_info_to_execute_invocation::<T>(
-                                    tx_exec_info
-                                        .execute_call_info
-                                        .as_ref()
-                                        .ok_or(Error::<T>::TransactionExecutionFailed)?,
-                                    tx_exec_info.revert_error.as_ref(),
-                                )?,
-                                fee_transfer_invocation,
-                                // TODO(#1291): Compute state diff correctly
-                                state_diff: Some(StateDiff::default()),
-                            })
-                        }
-                        UserTransaction::Declare(_, _) => {
-                            TransactionTrace::Declare(DeclareTransactionTrace {
-                                validate_invocation,
-                                fee_transfer_invocation,
-                                // TODO(#1291): Compute state diff correctly
-                                state_diff: Some(StateDiff::default()),
-                            })
-                        }
-                        UserTransaction::DeployAccount(_) => {
-                            TransactionTrace::DeployAccount(DeployAccountTransactionTrace {
-                                validate_invocation,
-                                constructor_invocation: tx_exec_info
-                                    .execute_call_info
-                                    .as_ref()
-                                    .ok_or(Error::<T>::TransactionExecutionFailed)?
-                                    .try_into()
-                                    .map_err(|_| Error::<T>::TransactionExecutionFailed)?,
-
-                                fee_transfer_invocation,
-                                // TODO(#1291): Compute state diff correctly
-                                state_diff: Some(StateDiff::default()),
-                            })
-                        }
-                    };
-                    let gas_consumed = tx_exec_info
-                        .execute_call_info
-                        .as_ref()
-                        .map(|x| x.execution.gas_consumed)
-                        .ok_or(Error::<T>::TransactionExecutionFailed)?;
-                    let overall_fee = tx_exec_info.actual_fee.0 as u64;
-                    let gas_price = if gas_consumed > 0 { overall_fee / gas_consumed } else { 0 };
-                    results.push(SimulatedTransaction {
-                        transaction_trace,
-                        fee_estimation: FeeEstimate { gas_consumed, gas_price, overall_fee },
-                    });
-                }
-                Err(_e) => {
-                    return Err(Error::<T>::TransactionExecutionFailed.into());
-                }
-            }
-        }
-
-        Ok(results)
-    }
-
     pub fn emit_and_store_tx_and_fees_events(
         tx_hash: TransactionHash,
         execute_call_info: &Option<CallInfo>,
@@ -1364,6 +1232,15 @@ impl<T: Config> Pallet<T> {
 
     pub fn program_hash() -> Felt252Wrapper {
         T::ProgramHash::get()
+    }
+
+    pub fn config_hash() -> StarkHash {
+        Felt252Wrapper::from(T::SystemHash::compute_hash_on_elements(&[
+            FieldElement::from_byte_slice_be(SN_OS_CONFIG_HASH_VERSION.as_bytes()).unwrap(),
+            T::ChainId::get().into(),
+            Felt252Wrapper::from(Self::fee_token_address().0.0).into(),
+        ]))
+        .into()
     }
 
     pub fn is_transaction_fee_disabled() -> bool {